Switched gear changes to short

+++ src/SessionDocumentation/Lap.cs
@@ -18,7 +18,7 @@ namespace ApexVisual.SessionDocumentation
         public byte PercentOnMaxThrottle {get; set;}
         public byte PercentOnMaxBrake {get; set;}
         public float EndingErs {get; set;}
-        public byte GearChanges {get; set;}
+        public short GearChanges {get; set;}
         public TyreCompound EquippedTyreCompound {get; set;}
         public Guid EndingTyreWear {get; set;}
